Move onCreatePage out into own file
