+++ package.json
@@ -51,6 +51,7 @@
     "gatsby-plugin-google-tagmanager": "^2.1.25",
     "gatsby-plugin-react-helmet": "^3.1.14",
     "gatsby-plugin-sitemap": "^2.2.27",
+    "memoizerific": "^1.11.3",
     "mongodb-stitch-browser-sdk": "^4.7.1",
     "mongodb-stitch-server-sdk": "^4.7.1",
     "prop-types": "^15.7.2",
